+++ src/TRegx/CleanRegex/Split/FilteredSplitPattern.php
@@ -1,11 +1,10 @@
 <?php
-namespace TRegx\CleanRegex;
+namespace TRegx\CleanRegex\Split;
 
 use TRegx\CleanRegex\Internal\InternalPattern;
-use TRegx\CleanRegex\Split\FilteredSplitPattern;
 use TRegx\SafeRegex\preg;
 
-class SplitPattern
+class FilteredSplitPattern
 {
     /** @var InternalPattern */
     private $pattern;
@@ -16,11 +15,6 @@ class SplitPattern
     {
         $this->pattern = $pattern;
         $this->subject = $subject;
-    }
-
-    public function filter(): FilteredSplitPattern
-    {
-        return new FilteredSplitPattern($this->pattern, $this->subject);
     }
 
     /**
@@ -42,6 +36,6 @@ class SplitPattern
     private function split(bool $includeDelimiter): array
     {
         $flag = $includeDelimiter ? PREG_SPLIT_DELIM_CAPTURE : 0;
-        return preg::split($this->pattern->pattern, $this->subject, -1, $flag);
+        return preg::split($this->pattern->pattern, $this->subject, -1, $flag | PREG_SPLIT_NO_EMPTY);
     }
 }